[IMP] mail: clean up make store
Move markRaw to constructor instead of new() caller as these classes
should always be raw regardless of the caller.

Move record proxy to record internal instead of bloating make store. The
proxy is record logic so it belongs to record internal,  make store is
more about bootstrapping and Models building.
